@@ -1,18 +1,22 @@
-import { Directive, HostBinding, Input } from '@angular/core'
+import { Directive, computed, input } from '@angular/core'
 
 @Directive({
   selector: '[projectStyle]',
+  host: {
+    '[style]': 'style()',
+  },
 })
 export class ProjectStyleDirective {
-  @Input({ required: true }) projectStyle!: number
+  readonly projectStyle = input.required<number>()
 
-  @HostBinding('style') get style() {
-    return `
+  readonly style = computed(
+    () =>
+      `
       border: 1px solid gray;
       border-radius: 5px;
       padding: 8px;
       font-size: 14px;
-      background: hsla(${this.projectStyle * 30}, 60%, 80%, 1);
-    `
-  }
+      background: hsla(${this.projectStyle() * 30}, 60%, 80%, 1);
+    `,
+  )
 }
