@@ -14,19 +14,22 @@ import { TasksService } from '../services/tasks.service'
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
   selector: 'auto-refetching-example',
-  standalone: true,
   templateUrl: './auto-refetching.component.html',
   imports: [NgStyle],
 })
 export class AutoRefetchingExampleComponent {
-  #tasksService = inject(TasksService)
+  readonly #tasksService = inject(TasksService)
 
-  intervalMs = signal(1000)
+  readonly intervalMs = signal(1000)
 
-  tasks = injectQuery(() => this.#tasksService.allTasks(this.intervalMs()))
+  readonly tasks = injectQuery(() =>
+    this.#tasksService.allTasks(this.intervalMs()),
+  )
 
-  addMutation = injectMutation(() => this.#tasksService.addTask())
-  clearMutation = injectMutation(() => this.#tasksService.clearAllTasks())
+  readonly addMutation = injectMutation(() => this.#tasksService.addTask())
+  readonly clearMutation = injectMutation(() =>
+    this.#tasksService.clearAllTasks(),
+  )
 
   clearTasks() {
     this.clearMutation.mutate()
