@@ -16,7 +16,7 @@ interface ProjectResponse {
   providedIn: 'root',
 })
 export class ProjectsService {
-  #http = inject(HttpClient)
+  readonly #http = inject(HttpClient)
   getProjects = (page: number) =>
     this.#http.get<ProjectResponse>(`/api/projects?cursor=${page}`)
 }
